Add files via upload
In an area of 1000sft, we need to fit the following requirement

Five rectangle of 80sft. 
One rectangle of 100 sft. 
One rectangle of 150 sft

Note:

We should be in position to change the required for the give area (1000sft). In scenario where is requirement exceeding the area, an error should pop up stating the same. No two areas are to intersect.